scrape_prices/src/requirements.txt
pandas==1.5.2
pandas-gbq==0.18.0
google-cloud-secret-manager==2.12.6
entsoe-py==0.5.8
xmltodict==0.13.0
